requirements.txt
web3>=6.0.0
eth-account>=0.8.0
py-solc-x>=1.12.0
colorama>=0.4.6

